requirements.txt
streamlit
pandas
numpy
joblib
groq
langchain-groq
python-dotenv
scikit-learn
xgboost
langchain